shairport-metadata-server/requirements.txt
flask>=2.0.0
paho-mqtt>=1.6.0
